requirements.txt
streamlit
pdfplumber
faiss-cpu
numpy
sentence-transformers
google-generativeai
reportlab
asyncio
aiohttp
